pyproject.toml
[project]
name = "tool-use-brittleness-ablation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "addict>=2.4.0",
    "deepdiff>=8.6.1",
    "docstring-parser>=0.17.0",
    "fs>=2.4.16",
    "gymnasium>=1.2.3",
    "loguru>=0.7.3",
    "openai>=2.21.0",
    "pydantic-argparse>=0.10.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "tiktoken>=0.12.0",
    "toml>=0.10.2",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
]
